pyproject.toml
[project]
name = "qsmec-command-center"
version = "2.0.0"
description = "Q-SMEC Command Center — unified operational hub for NIKET NA LLC"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "sqlmodel>=0.0.16",
    "sqlalchemy>=2.0",
    "pydantic>=2.5",
    "google-auth>=2.27",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.120",
    "pygithub>=2.3",
    "anthropic>=0.50",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "aiosqlite>=0.20",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "deptry>=0.25",
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.3",
    "pip-audit",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.deptry]
extend_exclude = [".venv", "tests", "docs", "build"]

[tool.deptry.package_module_name_map]
python-dotenv = "dotenv"
pygithub = "github"
google-api-python-client = "googleapiclient"
google-auth-oauthlib = "google_auth_oauthlib"
google-auth = "google"

[tool.deptry.per_rule_ignores]
DEP001 = ["backend", "client", "email_cache", "note", "pipeline", "task"]
DEP002 = ["uvicorn", "httpx", "google-genai", "jinja2", "aiosqlite", "deptry", "pytest", "ruff", "pip-audit"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
